Refactoring: Move whole application overview-tab into the app-detail-overview component
