+++ mico-admin/src/app/app-detail/app-detail.component.ts
@@ -23,11 +23,10 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { ApiService } from '../api/api.service';
 import { ApiObject } from '../api/apiobject';
 import { Subscription } from 'rxjs';
-import { MatDialog } from '@angular/material';
-import { ServicePickerComponent } from '../dialogs/service-picker/service-picker.component';
 import { versionComparator } from '../api/semantic-version';
 import { YesNoDialogComponent } from '../dialogs/yes-no-dialog/yes-no-dialog.component';
 import { CreateNextVersionComponent } from '../dialogs/create-next-version/create-next-version.component';
+import { MatDialog } from '@angular/material';
 
 @Component({
     selector: 'mico-app-detail',
@@ -39,18 +38,18 @@ export class AppDetailComponent implements OnInit, OnDestroy {
     constructor(
         private apiService: ApiService,
         private route: ActivatedRoute,
-        private dialog: MatDialog,
         private router: Router,
+        private dialog: MatDialog,
     ) { }
 
     subRouteParams: Subscription;
     subApplicationVersions: Subscription;
     subDeploy: Subscription;
-    subDependeesDialog: Subscription;
     subPublicIps: Subscription[] = [];
     subApplication: Subscription;
     subServiceDependency: Subscription;
     subCreateNextVersion: Subscription;
+
 
     // immutable application  object which is updated, when new data is pushed
     application: ApiObject;
@@ -147,7 +146,6 @@ export class AppDetailComponent implements OnInit, OnDestroy {
         this.unsubscribe(this.subRouteParams);
         this.unsubscribe(this.subApplicationVersions);
         this.unsubscribe(this.subDeploy);
-        this.unsubscribe(this.subDependeesDialog);
         this.subPublicIps.forEach(subscription => {
             this.unsubscribe(subscription);
         });
@@ -178,58 +176,6 @@ export class AppDetailComponent implements OnInit, OnDestroy {
         return this.allVersions[this.allVersions.length - 1].version;
     }
 
-    addService() {
-
-        const dialogRef = this.dialog.open(ServicePickerComponent, {
-            data: {
-                filter: '',
-                choice: 'multi',
-                existingDependencies: this.application.services,
-                serviceId: '',
-            }
-        });
-        this.subDependeesDialog = dialogRef.afterClosed().subscribe(result => {
-
-            if (result === '') {
-                return;
-            }
-
-            // TODO consider if null check is still neccesary as soon as endpoint to add dependencies exists
-            if (this.application.services == null) {
-                this.application.services = [];
-            }
-
-            result.forEach(service => {
-                // this.application.services.push(element);
-                // TODO Consider adding all at once.
-                this.apiService.postApplicationServices(this.application.shortName, this.application.version, service)
-                    .subscribe();
-            });
-        });
-
-    }
-
-    deleteService(serviceShortName: string) {
-
-        const dialogRef = this.dialog.open(YesNoDialogComponent, {
-            data: {
-                object: serviceShortName,
-                question: 'deleteDependency'
-            }
-        });
-
-        this.subServiceDependency = dialogRef.afterClosed().subscribe(shouldDelete => {
-            if (shouldDelete) {
-
-                this.apiService.deleteApplicationServices(this.application.shortName, this.application.version, serviceShortName)
-                    .subscribe(val => {
-                        // TODO add some user output (as soon as the endpoint actually exists)
-
-                    });
-            }
-        });
-
-    }
 
     /**
     * call-back from the version picker
